Fix build errors and warnings
- Fix critical compilation error: uncomment with_longterm_reid_correction parameter in hybridsort.cpp
- Fix initialization order warnings in xysr_kf.cpp, onnx_backend.cpp, and hybridsort.cpp
- Fix unused parameter warnings across multiple tracker files
- Improve CMakeLists.txt to conditionally link opencv_video and opencv_calib3d targets
- Fix sign comparison warnings in bytetrack.cpp and strongsort.cpp

All builds now pass successfully on Linux.

+++ src/trackers/bytetrack.cpp
@@ -69,7 +69,7 @@ void STrack::update(const STrack& new_track, int frame_id) {
     frame_id_ = frame_id;
     tracklet_len_ += 1;
     history_observations_.push_back(xyxy());
-    if (history_observations_.size() > max_obs_) {
+    if (static_cast<int>(history_observations_.size()) > max_obs_) {
         history_observations_.pop_front();
     }
